requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.6.2.post1
APScheduler==3.6.3
asgiref==3.8.1
attrs==25.3.0
audioop-lts==0.2.1
cachetools==4.2.2
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
discord.py==2.5.2
dj-database-url==2.3.0
Django==5.1.7
docutils==0.21.2
filetype==1.2.0
frozenlist==1.7.0
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
libretranslatepy==2.1.1
lxml==5.3.0
multidict==6.5.0
packaging==25.0
propcache==0.3.2
pygame==2.6.1
Pygments==2.19.1
pypiwin32==223
pytchat==0.5.5
pyTelegramBotAPI==4.24.0
python-decouple==3.8
python-dotenv==1.1.0
python-telegram-bot==13.15
pytz==2024.2
pywin32==308
requests==2.32.3
rfc3986==1.5.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.1
translate==3.6.1
typing_extensions==4.13.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
whitenoise==6.9.0
yarl==1.20.1
